+++ CHANGELOG.md
@@ -45,6 +45,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `GetAddressBuilder` and `AddressBalance`;
 
+### Removed
+
+- `offline` field in `ClientBuilder`;
+
 ### Fixed
 
 - Check expiration for remainder address;